pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ogg-summarizer"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "assemblyai>=0.40.2",
    "ipython>=9.2.0",
    "openai>=1.76.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "speechrecognition>=3.14.2",
]

[tool.setuptools]
packages = ["summarizer"]

[project.scripts]
summarize = "summarizer.main:main"
